+++ Code/src/main/java/com/huaDevelopers/data/Services/Interfaces/InsuranceService.java
@@ -12,13 +12,13 @@ public interface InsuranceService {
 
 	public void updateInsurance(Insurance insurance);
 
-	public Insurance getInsuranceByID(int id);
+	public Insurance getInsuranceByID(Long id);
 
 	public Insurance getInsuranceByLicensePlate(String licensePlate);
 
 	public List<Insurance> listAllInsurances();
 
-	public void removeInsurance(int id);
+	public void removeInsurance(Long id);
 
 	public double countInsurCost(Vehicle vehicle, Customer cust, String type, int duration, boolean flag);
